Extracting Agents Context
This context manager everything related to the two agents. Previously all this was generally available in the AppContext which was managing the general global app state. But I think now, with two separate agents and them both accessing the same history + making journal entries, it makes a lot more sense to have all this in a separate context.
Another reason I am doing it this way is because I want to eventually export this into a separate server that does all these experiences